pyproject.toml
[project]
name = "caatingaid"
version = "0.1.0"
description = ""
authors = [
    {name = "DeveloperKairo",email = "contatodevkairo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai (>=0.8.6,<0.9.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "caatingaid", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
